backend/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
dotenv==0.9.9
fastapi==0.117.1
faxplus-api @ git+https://github.com/alohi/faxplus-python.git@947370631cb1c7cb5d11a7d24d74e3ab64714740
h11==0.16.0
idna==3.10
pillow==11.3.0
pydantic==2.11.9
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
reportlab==4.4.4
requests==2.32.5
six==1.17.0
sniffio==1.3.1
starlette==0.48.0
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.36.0
